+++ docs/versionhistory.rst
@@ -9,6 +9,9 @@ APScheduler, see the :doc:`migration section <migration>`.
 
 * Added the ``engine_options`` option to ``SQLAlchemyJobStore``
 
+* Fixed memory leak due to a cyclic reference when jobs raise exceptions
+  (thanks to gilbsgilbs for help on solving this)
+
 
 3.4.0
 -----